+++ NEWS.md
@@ -1,3 +1,9 @@
+# serializer 0.1.2   2020-09-22
+
+* Feature: `calc_marshalled_size()` added. This function will only calculate
+  the size of the ersulting serialized object, but will not do any actual
+  writing of bytes to the output.
+
 # serializer 0.1.1   2020-09-22
 
 * use 'marshall/unmarshall' function names instead of 'pack/unpack'